refactor(app): extract chat store
